backend/pyproject.toml
[project]
name = "mimir-backend"
version = "0.1.0"
description = "Mimir — Self-hosted AI second brain agent"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "httpx>=0.27.0",
    "aiosqlite>=0.20.0",
    "chromadb>=0.5.0",
    "apscheduler>=3.10",
    "trafilatura>=1.12.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0",
    "bleach>=6.0",
    "pymupdf>=1.24.0",
    "python-multipart>=0.0.9",
    "watchdog>=4.0",
    "python-docx>=1.1.0",
    "aioimaplib>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
